Add UI for Courses day widget
