+++ Widget/WidgetBundle.swift
@@ -9,15 +9,9 @@
 import WidgetKit
 import SwiftUI
 
-struct EmptyWidget: Widget {
-    var body: some WidgetConfiguration {
-        EmptyWidgetConfiguration()
-    }
-}
-
 @main
 struct LabsWidgetBundle: WidgetBundle {
     var body: some Widget {
-        EmptyWidget()
+        CoursesDayWidget()
     }
 }
